Add logging mixin functionality
In this commit, the logging mixin functionality has been added to the application. The mixin, named loggingMixin, provides a logRequest function that logs the details of each incoming request, including the HTTP method and URL. The loggingMixin object has been assigned to the express app using Object.assign, allowing the app to utilize the mixin's functionality.